+++ loraham_daemon/daemon_radio_init.h
@@ -1,0 +1,8 @@
+#ifndef LORAHAM_DAEMON_RADIO_INIT_H
+#define LORAHAM_DAEMON_RADIO_INIT_H
+
+/* --- Radio startup/init -------------------------------------------------- */
+
+void lora_init(void);
+
+#endif